# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RFID Tool Tracking System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface RegisterBody {
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export type ToolStatus = (typeof ToolStatus)[keyof typeof ToolStatus];

export const ToolStatus = {
  Available: "Available",
  Issued: "Issued",
  Missing: "Missing",
} as const;

export interface Tool {
  id: number;
  toolId: string;
  name: string;
  category: string;
  status: ToolStatus;
  createdAt: string;
  updatedAt: string;
}

export type CreateToolBodyStatus =
  (typeof CreateToolBodyStatus)[keyof typeof CreateToolBodyStatus];

export const CreateToolBodyStatus = {
  Available: "Available",
  Issued: "Issued",
  Missing: "Missing",
} as const;

export interface CreateToolBody {
  toolId: string;
  name: string;
  category: string;
  status?: CreateToolBodyStatus;
}

export type UpdateToolBodyStatus =
  (typeof UpdateToolBodyStatus)[keyof typeof UpdateToolBodyStatus];

export const UpdateToolBodyStatus = {
  Available: "Available",
  Issued: "Issued",
  Missing: "Missing",
} as const;

export interface UpdateToolBody {
  name?: string;
  category?: string;
  status?: UpdateToolBodyStatus;
}

export interface IssueToolBody {
  toolId: string;
  userId: string;
}

export interface ReturnToolBody {
  toolId: string;
}

export type TransactionActionType =
  (typeof TransactionActionType)[keyof typeof TransactionActionType];

export const TransactionActionType = {
  issue: "issue",
  return: "return",
} as const;

export interface Transaction {
  id: number;
  toolId: string;
  /** @nullable */
  userId?: string | null;
  actionType: TransactionActionType;
  /** @nullable */
  issueDate?: string | null;
  /** @nullable */
  returnDate?: string | null;
  createdAt: string;
  /** @nullable */
  toolName?: string | null;
}

export interface ScanBody {
  scannedTools: string[];
}

export type ScanResultSummary = {
  total: number;
  correct: number;
  missing: number;
  extra: number;
};

export interface ScanResult {
  correctTools: string[];
  missingTools: string[];
  extraTools: string[];
  summary: ScanResultSummary;
}

export interface CategoryCount {
  category: string;
  count: number;
}

export interface Stats {
  totalTools: number;
  availableTools: number;
  issuedTools: number;
  missingTools: number;
  totalTransactions: number;
  recentTransactions: Transaction[];
  categoryBreakdown: CategoryCount[];
}

export type ListToolsParams = {
  status?: ListToolsStatus;
  category?: string;
};

export type ListToolsStatus =
  (typeof ListToolsStatus)[keyof typeof ListToolsStatus];

export const ListToolsStatus = {
  Available: "Available",
  Issued: "Issued",
  Missing: "Missing",
} as const;

export type ListTransactionsParams = {
  toolId?: string;
  limit?: number;
};