# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RFID Tool Tracking System API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Register a new user
 */
export const registerBodyPasswordMin = 6;

export const RegisterBody = zod.object({
  name: zod.string(),
  email: zod.string().email(),
  password: zod.string().min(registerBodyPasswordMin),
});

/**
 * @summary Login
 */
export const LoginBody = zod.object({
  email: zod.string().email(),
  password: zod.string(),
});

export const LoginResponse = zod.object({
  token: zod.string(),
  user: zod.object({
    id: zod.number(),
    name: zod.string(),
    email: zod.string(),
    createdAt: zod.coerce.date(),
  }),
});

/**
 * @summary List all tools
 */
export const ListToolsQueryParams = zod.object({
  status: zod.enum(["Available", "Issued", "Missing"]).optional(),
  category: zod.coerce.string().optional(),
});

export const ListToolsResponseItem = zod.object({
  id: zod.number(),
  toolId: zod.string(),
  name: zod.string(),
  category: zod.string(),
  status: zod.enum(["Available", "Issued", "Missing"]),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const ListToolsResponse = zod.array(ListToolsResponseItem);

/**
 * @summary Create a tool
 */
export const createToolBodyStatusDefault = `Available`;

export const CreateToolBody = zod.object({
  toolId: zod.string(),
  name: zod.string(),
  category: zod.string(),
  status: zod
    .enum(["Available", "Issued", "Missing"])
    .default(createToolBodyStatusDefault),
});

/**
 * @summary Get a tool by ID
 */
export const GetToolParams = zod.object({
  id: zod.coerce.number(),
});

export const GetToolResponse = zod.object({
  id: zod.number(),
  toolId: zod.string(),
  name: zod.string(),
  category: zod.string(),
  status: zod.enum(["Available", "Issued", "Missing"]),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Update a tool
 */
export const UpdateToolParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateToolBody = zod.object({
  name: zod.string().optional(),
  category: zod.string().optional(),
  status: zod.enum(["Available", "Issued", "Missing"]).optional(),
});

export const UpdateToolResponse = zod.object({
  id: zod.number(),
  toolId: zod.string(),
  name: zod.string(),
  category: zod.string(),
  status: zod.enum(["Available", "Issued", "Missing"]),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Delete a tool
 */
export const DeleteToolParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Issue a tool to a user
 */
export const IssueToolBody = zod.object({
  toolId: zod.string(),
  userId: zod.string(),
});

export const IssueToolResponse = zod.object({
  id: zod.number(),
  toolId: zod.string(),
  userId: zod.string().nullish(),
  actionType: zod.enum(["issue", "return"]),
  issueDate: zod.coerce.date().nullish(),
  returnDate: zod.coerce.date().nullish(),
  createdAt: zod.coerce.date(),
  toolName: zod.string().nullish(),
});

/**
 * @summary Return a tool
 */
export const ReturnToolBody = zod.object({
  toolId: zod.string(),
});

export const ReturnToolResponse = zod.object({
  id: zod.number(),
  toolId: zod.string(),
  userId: zod.string().nullish(),
  actionType: zod.enum(["issue", "return"]),
  issueDate: zod.coerce.date().nullish(),
  returnDate: zod.coerce.date().nullish(),
  createdAt: zod.coerce.date(),
  toolName: zod.string().nullish(),
});

/**
 * @summary List all transactions
 */
export const listTransactionsQueryLimitDefault = 50;

export const ListTransactionsQueryParams = zod.object({
  toolId: zod.coerce.string().optional(),
  limit: zod.coerce.number().default(listTransactionsQueryLimitDefault),
});

export const ListTransactionsResponseItem = zod.object({
  id: zod.number(),
  toolId: zod.string(),
  userId: zod.string().nullish(),
  actionType: zod.enum(["issue", "return"]),
  issueDate: zod.coerce.date().nullish(),
  returnDate: zod.coerce.date().nullish(),
  createdAt: zod.coerce.date(),
  toolName: zod.string().nullish(),
});
export const ListTransactionsResponse = zod.array(ListTransactionsResponseItem);

/**
 * @summary Perform RFID inventory scan
 */
export const InventoryScanBody = zod.object({
  scannedTools: zod.array(zod.string()),
});

export const InventoryScanResponse = zod.object({
  correctTools: zod.array(zod.string()),
  missingTools: zod.array(zod.string()),
  extraTools: zod.array(zod.string()),
  summary: zod.object({
    total: zod.number(),
    correct: zod.number(),
    missing: zod.number(),
    extra: zod.number(),
  }),
});

/**
 * @summary Get dashboard statistics
 */
export const GetStatsResponse = zod.object({
  totalTools: zod.number(),
  availableTools: zod.number(),
  issuedTools: zod.number(),
  missingTools: zod.number(),
  totalTransactions: zod.number(),
  recentTransactions: zod.array(
    zod.object({
      id: zod.number(),
      toolId: zod.string(),
      userId: zod.string().nullish(),
      actionType: zod.enum(["issue", "return"]),
      issueDate: zod.coerce.date().nullish(),
      returnDate: zod.coerce.date().nullish(),
      createdAt: zod.coerce.date(),
      toolName: zod.string().nullish(),
    }),
  ),
  categoryBreakdown: zod.array(
    zod.object({
      category: zod.string(),
      count: zod.number(),
    }),
  ),
});